:sparkles:Feat: Card, Tag 수정 및 테스트 완료

+++ src/main/java/com/likelion/trendithon/domain/tag/Service/TagService.java
@@ -1,27 +1,38 @@
 package com.likelion.trendithon.domain.tag.Service;
 
-import com.likelion.trendithon.domain.tag.entity.Tag;
-import com.likelion.trendithon.domain.tag.repository.TagRepository;
 import java.util.List;
 import java.util.Objects;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.likelion.trendithon.domain.card.entity.Card;
+import com.likelion.trendithon.domain.tag.dto.TagDto;
+import com.likelion.trendithon.domain.tag.entity.Tag;
+import com.likelion.trendithon.domain.tag.repository.TagRepository;
 
 import lombok.AllArgsConstructor;
-import org.springframework.transaction.annotation.Transactional;
+import lombok.extern.slf4j.Slf4j;
 
 @Service
 @AllArgsConstructor
+@Slf4j
 public class TagService {
 
   private TagRepository tagRepository;
 
-  @Transactional
-  public void createTag(List<Tag> tags) {
-
-    tagRepository.saveAll(tags);
+  public void saveTag(List<TagDto> tags, Card newCard) {
+    List<Tag> newTags =
+        tags.stream()
+            .map(
+                tagDto ->
+                    Tag.builder()
+                        .tagTitle(tagDto.getTagTitle())
+                        .tagContent(tagDto.getTagContent())
+                        .card(newCard)
+                        .build())
+            .toList();
+    tagRepository.saveAll(newTags);
   }
 
   @Transactional
@@ -31,35 +42,48 @@ public class TagService {
   }
 
   @Transactional
-  public List<Tag> updateTags(Card card) {
+  public void updateTags(Card card, List<TagDto> newTags) {
 
     List<Tag> tags = tagRepository.findByCard(card);
-    List<Tag> newTags = card.getTagItems();
 
     // tags엔 존재하는데 newTags에 존재하지 않으면 삭제 같으면 업뎃
     for (Tag tag : tags) {
-      for (Tag newTag : newTags) {
-        if (Objects.equals(tag.getTagId(), newTag.getTagId())) {
+      boolean found = false;
+      for (TagDto newTag : newTags) {
+        if (Objects.equals(tag.getTagTitle(), newTag.getTagTitle())) {
           tag.setTagTitle(newTag.getTagTitle());
           tag.setTagContent(newTag.getTagContent());
+          found = true;
+        }
+      }
+
+      if (!found) {
+        tagRepository.delete(tag);
+      }
+    }
+
+    // newTags엔 존재하는데 tags엔 존재하지 않으면 추가
+    for (TagDto newTag : newTags) {
+      // 기존 [1,2,3,4] 업데이트 [2,3,6]
+      boolean found = false;
+
+      for (Tag tag : tags) {
+        if (Objects.equals(tag.getTagTitle(), newTag.getTagTitle())) {
+          found = true;
           break;
         }
       }
 
-      tags.remove(tag);
-      tagRepository.delete(tag);
-    }
+      if (!found) {
 
-    // newTags엔 존재하는데 tags엔 존재하지 않으면 추가
-    for (Tag newTag : newTags) {
-      for (Tag tag : tags) {
-        if (Objects.equals(tag.getTagId(), newTag.getTagId())) break;
+        Tag tmp =
+            Tag.builder()
+                .tagTitle(newTag.getTagTitle())
+                .tagContent(newTag.getTagContent())
+                .card(card)
+                .build();
+        tagRepository.save(tmp);
       }
-
-      tags.add(newTag);
-      tagRepository.save(newTag);
     }
-
-    return tags;
   }
 }
